# Conflicts:
#	CMakeLists.txt
#	src/App.cpp
#	src/Triangle.cpp

+++ include/Core/IndexBuffer.hpp
@@ -19,7 +19,7 @@ public:
     IndexBuffer &operator=(const IndexBuffer &) = delete;
     IndexBuffer &operator=(IndexBuffer &&other);
 
-    unsigned int GetCount() const { return m_Count; }
+    size_t GetCount() const { return m_Count; }
 
     void Bind() const;
     void Unbind() const;
@@ -27,7 +27,7 @@ public:
 private:
     GLuint m_BufferId;
 
-    unsigned int m_Count;
+    size_t m_Count;
 };
 } // namespace Core
